CONSOLE-4580: use typescript Record type for the key value list of ResourceActionCreators, replace deprecated K8sKind type, move the ResourceActionCreator ResourceActionFactory to separate types file.
