+++ frontend/packages/console-app/src/actions/creators/health-checks-factory.ts
@@ -2,7 +2,7 @@ import i18next from 'i18next';
 import * as _ from 'lodash';
 import { Action } from '@console/dynamic-plugin-sdk';
 import { K8sResourceKind, K8sKind, referenceFor } from '@console/internal/module/k8s';
-import { ResourceActionFactory } from './common-factory';
+import { ResourceActionFactory } from './types';
 
 const healthChecksAdded = (resource: K8sResourceKind): boolean => {
   const containers = resource?.spec?.template?.spec?.containers;
