+++ frontend/packages/console-app/src/actions/creators/job-factory.ts
@@ -1,7 +1,7 @@
 import i18next from 'i18next';
 import { configureJobParallelismModal } from '@console/internal/components/modals';
 import { JobKind, K8sKind } from '@console/internal/module/k8s';
-import { ResourceActionFactory } from './common-factory';
+import { ResourceActionFactory } from './types';
 
 export const JobActionFactory: ResourceActionFactory = {
   ModifyJobParallelism: (kind: K8sKind, obj: JobKind) => ({
